+++ apps/doc/src/app/components/tree/tree.module.ts
@@ -3,7 +3,12 @@ import { CommonModule } from '@angular/common';
 import { PrizmAddonDocModule, prizmDocGenerateRoutes } from '@prizm-ui/doc';
 import { RouterModule } from '@angular/router';
 import { TreeComponent } from './tree.component';
-import { PrizmIconModule, PrizmTreeModule } from '@prizm-ui/components';
+import {
+  PrizmButtonModule,
+  PrizmIconModule,
+  PrizmSidebarModule,
+  PrizmTreeModule,
+} from '@prizm-ui/components';
 import { TreeBaseExampleComponent } from './examples/base/tree-base-example.component';
 import { TreeArrayExampleComponent } from './examples/array/tree-array-example.component';
 import { TreeTemplateExampleComponent } from './examples/template/tree-template-example.component';
@@ -12,6 +17,7 @@ import { TreeComponentExampleComponent } from './examples/component/tree-compone
 import { TreeTemplateCheckboxExampleModule } from './examples/template-checkbox/tree-template-example.module';
 import { TreeTemplateLazyExampleModule } from './examples/lazy/tree-template-lazy-example.module';
 import { TreePaddingInputExampleComponent } from './examples/padding-indent/tree-padding-indent-example.component';
+import { TreeInModalExampleComponent } from './examples/in-modal/tree-in-modal-example.component';
 
 @NgModule({
   imports: [
@@ -21,9 +27,12 @@ import { TreePaddingInputExampleComponent } from './examples/padding-indent/tree
     TreeTemplateLazyExampleModule,
     TreeTemplateCheckboxExampleModule,
     PrizmIconModule,
+    PrizmButtonModule,
+    PrizmSidebarModule,
     RouterModule.forChild(prizmDocGenerateRoutes(TreeComponent)),
   ],
   declarations: [
+    TreeInModalExampleComponent,
     TreeBaseExampleComponent,
     TreeArrayExampleComponent,
     TreeTemplateExampleComponent,